Fix deleting error

+++ services/propertyService.js
@@ -107,7 +107,9 @@ async function bid(id, userId, newPrice) {
 async function deleteProperty(id) {
     const property = Property.findById(id);
 
-    property.reviews.forEach(r => Review.findByIdAndDelete(r));
+    if (property.reviews?.length > 0) {
+        property.reviews.forEach(r => Review.findByIdAndDelete(r));
+    }
     return Property.findByIdAndDelete(id);
 }
